refactor(workspace): split module roots
- Move implementations out of lib.rs and mod.rs files
- Rename root test modules to responsibility-based files
