+++ crates/aion-agent/src/commands/registry_test.rs
