+++ crates/aion-agent/src/output/formatter_test.rs
