+++ crates/aion-mcp/src/transport/mod.rs
@@ -1,42 +1,6 @@
 pub mod sse;
 pub mod stdio;
 pub mod streamable_http;
+mod types;
 
-use async_trait::async_trait;
-
-use crate::protocol::{JsonRpcRequest, JsonRpcResponse};
-
-/// Transport abstraction for MCP communication
-#[async_trait]
-pub trait McpTransport: Send + Sync {
-    /// Send a JSON-RPC request and receive the response
-    async fn request(&self, req: &JsonRpcRequest) -> Result<JsonRpcResponse, McpError>;
-
-    /// Send a notification (no response expected)
-    async fn notify(&self, req: &JsonRpcRequest) -> Result<(), McpError>;
-
-    /// Close the transport
-    async fn close(&self) -> Result<(), McpError>;
-}
-
-/// Errors from MCP transport and protocol
-#[derive(Debug, thiserror::Error)]
-pub enum McpError {
-    #[error("Transport error: {0}")]
-    Transport(String),
-
-    #[error("JSON-RPC error {code}: {message}")]
-    JsonRpc { code: i64, message: String },
-
-    #[error("Server not found: {0}")]
-    ServerNotFound(String),
-
-    #[error("Tool not found: {server}/{tool}")]
-    ToolNotFound { server: String, tool: String },
-
-    #[error("Initialization failed: {0}")]
-    InitFailed(String),
-
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-}
+pub use types::{McpError, McpTransport};
