+++ crates/aion-skills/src/bundled/registry_test.rs
